pyproject.toml
[project]
name = "camera-scanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dlib>=20.0.0",
    "face-recognition>=1.3.0",
    "face-recognition-models",
    "flask>=3.1.3",
    "mediapipe<0.10.15",
    "numpy>=2.4.2",
    "onvif-zeep>=0.2.12",
    "opencv-python>=4.13.0.92",
    "pyqt6>=6.10.2",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.4.0",
    "pandas>=2.2.0",
    "qrcode[pil]>=8.2",
    "scapy>=2.7.0",
    "setuptools<70.0.0",
    "sounddevice>=0.5.5",
    "reolink-aio>=0.19.1",
]

[tool.uv.sources]
face-recognition-models = { git = "https://github.com/ageitgey/face_recognition_models.git" }
